Check has set cache driver in database or not and Add artisan command instead of hardcode

+++ database/migrations/create_cache_table.php
@@ -1,11 +1,13 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
+
+    const ARTISAN_CACHE_TABLE_CREATION = 'php artisan cache:table';
+
     /**
      * Run the migrations.
      *
@@ -13,18 +15,11 @@ return new class extends Migration
      */
     public function up()
     {
-        if (!Schema::hasTable('cache')) {
-            Schema::create('cache', function (Blueprint $table) {
-                $table->string('key')->primary();
-                $table->mediumText('value');
-                $table->integer('expiration');
-            });
-
-            Schema::create('cache_locks', function (Blueprint $table) {
-                $table->string('key')->primary();
-                $table->string('owner');
-                $table->integer('expiration');
-            });
+        if (
+            !Schema::hasTable('cache')
+            && config('acl.cache.storage') === 'database'
+        ) {
+            Artisan::call(static::ARTISAN_CACHE_TABLE_CREATION);
         }
     }
 
@@ -35,7 +30,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('cache');
-        Schema::dropIfExists('cache_locks');
+        if (config('acl.cache.storage') === 'database') {
+            Schema::dropIfExists('cache');
+            Schema::dropIfExists('cache_locks');
+        }
     }
 };
